(refactor): Optimización de lógica mejoras en velocidad de obtención de datos
