+++ src/renderer/src/hooks/Favorites/useFavoritesEffects.ts
@@ -1,12 +1,41 @@
-import { SymbolInfo } from '@renderer/interfaces/binance.interface';
-import { fetchAvailableSymbols, fetchAvailableSymbolsNoFilter } from '@renderer/services/binance-api.services';
-import { useEffect, useCallback } from 'react';
+import { fetchAvailableSymbolsNoFilter } from '@renderer/services/binance-api.services';
+import { useEffect, useCallback, useState, useMemo } from 'react';
+import { useFavoritesState } from './useFavoritesState';
 
 export const useFavoritesEffects = (
-  setSymbols: React.Dispatch<React.SetStateAction<SymbolInfo[]>>,
-  setLoading: React.Dispatch<React.SetStateAction<boolean>>
+    state: ReturnType<typeof useFavoritesState>
 ) => {
-  // Usamos useCallback para que la función sea estable
+
+  const {
+    setLoading,
+    setSymbols,
+    setFavorites,
+    favorites
+  }= state;
+
+  const [currentPage, setCurrentPage] = useState(1);
+  const itemsPerPage = 25;
+
+  const toggleFavorite = (symbol: string) => {
+    setFavorites(prev => 
+      prev.includes(symbol) ? prev.filter(s => s !== symbol) : [...prev, symbol]
+    );
+  };  
+
+  const filteredSymbols = useMemo(() => {
+    return state.symbols
+      .filter(item => favorites.includes(item.symbol)) // <--- CLAVE: Filtra solo lo que está en favoritos
+      .filter(item => 
+        item.symbol.toLowerCase().includes(state.search.toLowerCase()) ||
+        item.baseAsset.toLowerCase().includes(state.search.toLowerCase())
+      );
+  }, [state.symbols, state.search, favorites]); // Se añade 'favorites' como dependencia para refrescar la vista en tiempo real
+  const paginatedSymbols = useMemo(() => {
+    const start = (currentPage - 1) * itemsPerPage;
+    return filteredSymbols.slice(start, start + itemsPerPage);
+  }, [filteredSymbols, currentPage]);
+
+  const totalPages = Math.ceil(filteredSymbols.length / itemsPerPage);
   const loadSymbols = useCallback(async () => {
     setLoading(true);
     try {
@@ -18,10 +47,21 @@ export const useFavoritesEffects = (
       setLoading(false);
     }
   }, [setSymbols, setLoading]);
-
+  useEffect(() => {
+    setCurrentPage(1);
+  }, [state.search]);
+  useEffect(() => {
+    localStorage.setItem("crypto_favorites", JSON.stringify(favorites));
+  }, [favorites]);
+  
   useEffect(() => {
     loadSymbols();
   }, [loadSymbols]);
 
-  return { loadSymbols }; // Retornamos la función para usarla en el Dashboard
+  return { loadSymbols,
+    paginatedSymbols, 
+    toggleFavorite,   
+    currentPage,
+    setCurrentPage,
+    totalPages,}; // Retornamos la función para usarla en el Dashboard
 };
