+++ src/renderer/src/hooks/Favorites/useFavoritesState.ts
@@ -7,11 +7,21 @@ export const useFavoritesState = () => {
   const [loading, setLoading] = useState<boolean>(true);
   const [selectedSymbol, setSelectedSymbol] = useState<string | null>(null);
   const [search, setSearch] = useState<string>('');
+  const [favorites, setFavorites] = useState<string[]>(() => {
+    try {
+      const saved = localStorage.getItem("crypto_favorites");
+      return saved ? JSON.parse(saved) : [];
+    } catch (error) {
+      console.error("Error al recuperar favoritos:", error);
+      return [];
+    }
+  });
 
   return {
     symbols, setSymbols,
     loading, setLoading,
     selectedSymbol, setSelectedSymbol,
-    search, setSearch
+    search, setSearch,
+    favorites,setFavorites
   };
 };
