+++ src/renderer/src/styles/shared/AnalysisView.module.css
@@ -4,7 +4,9 @@
   padding: 10px;
   width: 100%;
   color: #eaecef;
-  height: 100%;
+  height: max-content;
+  overflow: auto;
+
 }
 
 .title {
