+++ src/renderer/src/styles/shared/Pagination.module.css
@@ -2,9 +2,10 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  gap: 15px;
-  margin-top: 20px;
-  padding: 15px;
+  gap: 10px;
+  border-radius: 10px;
+  margin-top: 10px;
+  height: 10%;
   background-color: #1e2329;
   border-top: 1px solid #2b3139;
   color: #eaecef;
